connected user and posts in db

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require("path");
 const { validationResult } = require("express-validator/check");
 
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -45,21 +46,31 @@ exports.createPost = (req, res, next) => {
     throw error;
   }
   const imageUrl = req.file.path;
-  const title = req.body.title;
-  const content = req.body.content;
+  const { title, content } = req.body;
+  let creator;
+
   const post = new Post({
     title,
     imageUrl,
     content,
-    creator: { name: "Mike" }
+    creator: req.userId
   });
 
   post
     .save()
     .then(result => {
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then(result => {
       res.status(201).json({
         message: "Post created succesfully!",
-        post: result
+        post,
+        creator: { _id: creator._id, name: creator.name }
       });
     })
     .catch(err => {
@@ -103,6 +114,7 @@ exports.updatePost = (req, res, next) => {
 
   const { title, content } = req.body;
   let imageUrl = req.body.image;
+
   if (req.file) {
     imageUrl = req.file.path;
   }
@@ -117,6 +129,11 @@ exports.updatePost = (req, res, next) => {
       if (!post) {
         const error = new Error("Post not found.");
         error.statusCode = 404;
+        throw error;
+      }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
         throw error;
       }
       if (imageUrl !== post.imageUrl) {
@@ -147,11 +164,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized.");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then(result => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then(user => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then(result => {
       res.status(200).json({ message: `Post ${postId} deleted` });
     })
     .catch(err => {
